backend/requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-import-export==3.2.0
django-utils-six==2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
psycopg2-binary==2.9.6
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
webcolors==1.13
xlrd==2.0.1
xlwt==1.3.0
